Storybook separate folder
