+++ src/stories/assets/heart.png
